% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_phylo.R
\name{boot_ci}
\alias{boot_ci}
\title{Generic method to output bootstrap confidence intervals from an object.}
\usage{
boot_ci(mod, ...)
}
\arguments{
\item{mod}{A \code{cor_phylo} object.}

\item{...}{Additional arguments.}
}
\value{
A list of confidence intervals.
}
\description{
Implemented only for \code{cor_phylo} objects thus far.
}
