% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{comm_a}
\alias{comm_a}
\title{Example community data}
\format{
A data frame with 15 sites and 15 species.
}
\usage{
comm_a
}
\description{
A data frame with site names as row names, species names as column names,
cells are the abundance of each species at each site.
}
\keyword{datasets}
