% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cor_phylo_}
\alias{cor_phylo_}
\title{Inner function to create necessary matrices and do model fitting.}
\usage{
cor_phylo_(X, U, SeM, Vphy_, REML, constrain_d, verbose, max_iter, method)
}
\arguments{
\item{X}{a n x p matrix with p columns containing the values for the n taxa.}

\item{U}{a list of p matrices corresponding to the p columns of X, with each
matrix containing independent variables for the corresponding column of X.}

\item{SeM}{a n x p matrix with p columns containing standard errors of the trait
values in X.}

\item{Vphy_}{}

\item{REML}{whether REML or ML is used for model fitting.}

\item{constrain_d}{if constrain.d is TRUE, the estimates of d are constrained to
be between zero and 1. This can make estimation more stable and can be
tried if convergence is problematic. This does not necessarily lead to
loss of generality of the results, because before using corphylo, branch
lengths of phy can be transformed so that the "starter" tree has strong
phylogenetic signal.}

\item{verbose}{if TRUE, the model logLik and running estimates of the correlation
coefficients and values of d are printed each iteration during optimization.}

\item{max_iter}{the maximum number of iterations in the optimization.}

\item{method}{method of optimization using nlopt. Options include
"bobyqa", "cobyla", "praxis". See
\url{https://nlopt.readthedocs.io/en/latest/NLopt_Algorithms/} for more
information.}
}
\value{
List containing output information, to be coerced to a cor_phylo object.
}
\description{
Inner function to create necessary matrices and do model fitting.
}
