% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{envi}
\alias{envi}
\title{Example environmental data}
\format{
A data frame with 15 sites and 4 variables: sand proportion,
canopy shade proportion, precipitation, and minimum temperature.
}
\usage{
envi
}
\description{
A data frame of site environmental variables.
}
\keyword{datasets}
