% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pglmm-utils.R
\name{family.communityPGLMM}
\alias{family.communityPGLMM}
\title{Family Objects for communityPGLMM objects}
\usage{
\method{family}{communityPGLMM}(object, ...)
}
\arguments{
\item{object}{the function \code{family} accesses the \code{family}
    objects which are stored within objects created by modelling
    functions (e.g., \code{glm}).}

\item{...}{further arguments passed to methods.}
}
\description{
Family Objects for communityPGLMM objects
}
