% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pglmm-utils.R
\name{fitted.communityPGLMM}
\alias{fitted.communityPGLMM}
\title{Fitted values for communityPGLMM}
\usage{
\method{fitted}{communityPGLMM}(object, ...)
}
\arguments{
\item{object}{a fitted model with class communityPGLMM.}

\item{\dots}{additional arguments, ignored for method compatibility.}
}
\value{
fitted values. For binomial PGLMMs, this is equal to mu (i.e. between 0 and 1).
}
\description{
Fitted values for communityPGLMM
}
