% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pglmm-utils.R
\name{fixed_effects.communityPGLMM}
\alias{fixed_effects.communityPGLMM}
\title{Extract coefficience for fixed terms}
\usage{
\method{fixed_effects}{communityPGLMM}(x)
}
\description{
Extract coefficience for fixed terms
}
