% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pglmm-utils.R
\name{fixed_effects}
\alias{fixed_effects}
\title{Extract coefficience for fixed terms}
\usage{
fixed_effects(x, ...)
}
\arguments{
\item{x}{A fitted model with class communityPGLMM.}

\item{...}{ignored.}
}
\value{
a dataframe of fixed-effects estimates.
}
\description{
Extract coefficience for fixed terms
}
