% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pglmm-utils.R
\name{fixef.communityPGLMM}
\alias{fixef.communityPGLMM}
\title{Extract coefficience for fixed terms}
\usage{
\method{fixef}{communityPGLMM}(object, ...)
}
\value{
a named, numeric vector of fixed-effects estimates.
}
\description{
Extract coefficience for fixed terms
}
