% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pglmm-utils.R
\docType{methods}
\name{fixef}
\alias{fixef}
\alias{fixef.communityPGLMM}
\alias{fixed.effects}
\title{Extract fixed-effects estimates}
\usage{
\method{fixef}{communityPGLMM}(object, ...)
}
\arguments{
\item{object}{A fitted model with class communityPGLMM.}

\item{...}{Ignored.}
}
\value{
A dataframe of fixed-effects estimates.
}
\description{
Extract the fixed-effects estimates
}
\details{
Extract the estimates of the fixed-effects parameters from a fitted model.
For bayesian models, the p-values are simply to indicate whether the
credible intervals include 0 (p = 0.04) or not (p = 0.6).
}
