% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{match_comm_tree}
\alias{match_comm_tree}
\title{Match phylogeny with community data}
\usage{
match_comm_tree(comm, tree, comm_2 = NULL)
}
\arguments{
\item{comm}{A site by species data frame, with site names as row names.}

\item{tree}{A phylogeny with "phylo" as class.}

\item{comm_2}{Another optional site by species data frame, if presented, both community data and the phylogeny
will have the same set of species. This can be useful for PCD with custom species pool.}
}
\value{
A list of the community data and the phylogeny.
}
\description{
This function will remove species from community data that are not in the phylogeny.
It will also remove tips from the phylogeny that are not in the community data.
}
