% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pglmm-utils.R
\name{nobs.communityPGLMM}
\alias{nobs.communityPGLMM}
\title{Number of Observation in a communityPGLMM Model}
\usage{
\method{nobs}{communityPGLMM}(object, use.fallback = FALSE, ...)
}
\arguments{
\item{object}{A fitted model object.}

\item{use.fallback}{logical: should fallback methods be used to try to
    guess the value?}

\item{...}{Further arguments to be passed to methods.}
}
\description{
Number of Observation in a communityPGLMM Model
}
