% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcd.R
\name{pcd}
\alias{pcd}
\title{pairwise site phylogenetic community dissimilarity (PCD) within a community}
\usage{
pcd(comm, tree, expectation = NULL, cpp = TRUE, verbose = TRUE)
}
\arguments{
\item{comm}{site by species data frame or matrix, sites as rows.}

\item{tree}{a phylogeny for species}

\item{expectation}{nsp_pool, psv_bar, psv_pool, and nsr calculated from \code{pcd_pred()}.}

\item{cpp}{whether to use loops written with c++, default is TRUE}

\item{verbose}{do you want to see the progress?}
}
\value{
a list of a variety of pairwise dissimilarities.
}
\description{
Calculate pairwise site PCD, users can specify expected values from \code{pcd_pred()}.
}
\examples{
x1 = pcd_pred(comm_1 = comm_a, comm_2 = comm_b, tree = tree, reps = 100)
pcd(comm = comm_a,
      tree = tree,
      expectation = x1)
}
