% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pglmm-plot.R
\name{plot.communityPGLMM}
\alias{plot.communityPGLMM}
\title{Plot the original dataset and predicted values (optional)}
\usage{
\method{plot}{communityPGLMM}(x, digits = max(3, getOption("digits") - 3),
  show.sp.names = FALSE, show.site.names = FALSE, predicted = FALSE, ...)
}
\arguments{
\item{x}{a fitted model with class communityPGLMM}

\item{digits}{minimal number of significant digits for printing, as in \code{\link{print.default}}}

\item{predicted}{whether to plot predicted values side by side with observed ones}

\item{...}{additional arguments for \code{Matrix::image()} or \code{lattice::levelplot()}.
Common ones are:
\itemize{
\item \code{useAbs} whether to use absolute values of the matrix; if no negative values,
this will be set to TRUE if not specified. When \code{useAbs = TRUE} the color scheme
will be black-white, otherwise, it will be red/blue.
\item \code{colorkey} whether to draw the scale legend at the right side of each plot?
}}

\item{show.sp.name}{whether to print species names as y-axis labels}

\item{show.site.name}{whether to print site names as x-axis labels}
}
\description{
Plot the original dataset and predicted values (optional)
}
\note{
the underlying plot grid object is returned but invisible. It can be saved for later uses.
}
