% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pglmm-utils.R
\name{communityPGLMM.predicted.values}
\alias{communityPGLMM.predicted.values}
\alias{pglmm.predicted.values}
\title{Predicted values of PGLMM}
\usage{
communityPGLMM.predicted.values(
  x,
  cpp = TRUE,
  gaussian.pred = c("nearest_node", "tip_rm")
)

pglmm.predicted.values(
  x,
  cpp = TRUE,
  gaussian.pred = c("nearest_node", "tip_rm")
)
}
\arguments{
\item{x}{A fitted model with class communityPGLMM.}

\item{cpp}{Whether to use c++ code. Default is TRUE.}

\item{gaussian.pred}{When family is gaussian, which type of prediction to calculate?
Option nearest_node will predict values to the nearest node, which is same as lme4::predict or
fitted. Option tip_rm will remove the point then predict the value of this point with remaining ones.}
}
\value{
A data frame with three columns: Y_hat (predicted values accounting for
both fixed and random terms), sp, and site.
}
\description{
\code{communityPGLMM.predicted.values} calculates the predicted
values of Y; for the generalized linear mixed model (family \%in\%
c("binomial","poisson"), these values are in the transformed space.
}
