% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pglmm-utils.R
\name{communityPGLMM.profile.LRT}
\alias{communityPGLMM.profile.LRT}
\alias{pglmm.profile.LRT}
\title{\code{communityPGLMM.profile.LRT} tests statistical significance of the
phylogenetic random effect of binomial models on
species slopes using a likelihood ratio test.}
\usage{
communityPGLMM.profile.LRT(x, re.number = 0, cpp = TRUE)

pglmm.profile.LRT(x, re.number = 0, cpp = TRUE)
}
\arguments{
\item{x}{A fitted model with class communityPGLMM and family "binomial".}

\item{re.number}{Which random term to test? Can be a vector with length >1}

\item{cpp}{Whether to use c++ function for optim. Default is TRUE. Ignored if \code{bayes = TRUE}.}
}
\value{
A list of likelihood, df, and p-value.
}
\description{
\code{communityPGLMM.profile.LRT} tests statistical significance of the
phylogenetic random effect of binomial models on
species slopes using a likelihood ratio test.
}
