% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pglmm-utils.R
\name{communityPGLMM.binary.LRT}
\alias{communityPGLMM.binary.LRT}
\alias{communityPGLMM.matrix.structure}
\alias{summary.communityPGLMM}
\alias{print.communityPGLMM}
\title{\code{communityPGLMM.binary.LRT} tests statistical significance of the phylogenetic random effect on
species slopes using a likelihood ratio test}
\usage{
communityPGLMM.binary.LRT(x, re.number = 0, cpp = TRUE)

communityPGLMM.matrix.structure(formula, data = list(), family = "binomial",
  tree, repulsion = FALSE, ss = 1, cpp = TRUE)

\method{summary}{communityPGLMM}(object, digits = max(3, getOption("digits") -
  3), ...)

\method{print}{communityPGLMM}(x, digits = max(3, getOption("digits") - 3),
  ...)
}
\arguments{
\item{x}{a fitted model with class communityPGLMM}

\item{re.number}{which random term to test? Can be a vector with length >1}

\item{cpp}{whether to use c++ function for optim. Default is TRUE. Ignored if \code{bayes = TRUE}.}

\item{formula}{a two-sided linear formula object describing the
mixed-effects of the model; it follows similar syntax with \code{\link[lme4:lmer]{lmer}}.
There are some differences though.

First, to specify that a random term should have phylogenetic cov matrix along
with non-phylogenetic one, add \code{__} (two underscores) at the end of the group variable,
e.g. \code{+ (1 | sp__)} will construct two random terms,
one with phylogenetic cov matrix and another with non-phylogenetic (Identity) matrix;
However, \code{__} in the nested terms (below) will only create a phlylogenetic cov-matrix.
Therefore, nested random term has four forms:
\enumerate{
\item \code{(1|sp__@site)} represents correlated species are nested within independent sites
(i.e. kronecker(I_sites, V_sp)). This should be the most common one for community analysis.
\item \code{(1|sp@site__)} represents independent species are nested within correlated sites
(i.e. kron(V_sites, I_sp)). This one can be used for bipartite questions.
You can, for example, treat sp as insects and site as plants.
Remember to set the argument \code{tree_site} to a phylogeny.
\item \code{(1|sp__@site__)} represents correlated species are nested within correlated sites
(i.e. kron(V_sites, V_sp)). This one can also be used for bipartite questions.
\item \code{(1|sp@site)} will generate a identity matrix, which will be the same as
the residual of LMM. So not meaningful.
}

Second, note that correlated random terms will not be allowed at this moment. For example,
\code{(x|g)} will be equal with \code{(0 + x|g)} in the \code{lme4::lmer} syntax;
also, \code{(x1 + x2|g)} won't work.}

\item{data}{a \code{\link{data.frame}} containing the variables
named in formula. The data frame should have long format with
a column for species (named as 'sp') and a column for sites (named as 'site').
\code{communityPGLMM} will reorder rows of the data frame so that species
are nested within sites (i.e. arrange first by column site then by column sp).}

\item{family}{either \code{gaussian} for a Linear Mixed Model, or
\code{binomial} for binary dependent data. If \code{bayes = TRUE}, \code{poisson} is also
supported.}

\item{tree}{a phylogeny for column sp, with "phylo" class. Or a var-cov matrix for sp,
make sure to have all species in the matrix; if the matrix is not standarized,
i.e. det(tree) != 1, we will try to standarize it for you.}

\item{repulsion}{when nested random term specified, do you want to test repulsion
(i.e. overdispersion) or underdispersion? Default is FALSE, i.e. test underdispersion.
This argument can be either a logical vector of length 1 or >1.
If its length is 1, then all cov matrices in nested terms will be either inverted (overdispersion) or not.
If its length is >1, then this means the users can select which cov matrix in the nested terms to be inverted.
If so, make sure to get the length right: for all the terms with \code{@},
count the number of "__" and this will be the length of repulsion.
For example, \code{sp__@site} will take one length as well as \code{sp@site__}.
\code{sp__@site__} will take two elements. So, if you nested terms are
\code{(1|sp__@site) + (1|sp@site__) + (1|sp__@site__)}
in the formula, then you should set the repulsion to be something like
\code{c(TRUE, FALSE, TURE, TURE)} (length of 4).
The T/F combinations depend on your questions.}

\item{ss}{which of the \code{random.effects} to produce}

\item{object}{a fitted model with class communityPGLMM.}

\item{digits}{minimal number of significant digits for printing, as in \code{\link{print.default}}}

\item{...}{additional arguments, currently ignored.}
}
\description{
\code{communityPGLMM.binary.LRT} tests statistical significance of the phylogenetic random effect on
species slopes using a likelihood ratio test

\code{communityPGLMM.matrix.structure} produces the entire
covariance matrix structure (V) when you specify random effects.
}
