% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pglmm-plot.R
\name{plot_bayes}
\alias{plot_bayes}
\title{plot_bayes generic}
\usage{
plot_bayes(x, ...)
}
\arguments{
\item{x}{A communityPGLMM object fit with \code{bayes = TRUE}.}

\item{...}{Further arguments to pass to or from other methods.}
}
\value{
A ggplot object
}
\description{
plot_bayes generic
}
