% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pglmm-utils.R
\name{predict.communityPGLMM}
\alias{predict.communityPGLMM}
\title{Predict Function for communityPGLMM Model Objects}
\usage{
\method{predict}{communityPGLMM}(object, ...)
}
\arguments{
\item{object}{a model object for which prediction is desired.}

\item{...}{additional arguments affecting the predictions produced.}
}
\value{

}
\description{
Predict Function for communityPGLMM Model Objects
}
