% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pglmm-utils.R
\name{print.communityPGLMM}
\alias{print.communityPGLMM}
\title{Print summary information of fitted model}
\usage{
\method{print}{communityPGLMM}(x, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{x}{A fitted communityPGLMM model.}

\item{digits}{Minimal number of significant digits for printing, as in \code{\link{print.default}}.}

\item{...}{Additional arguments, currently ignored.}
}
\description{
Print summary information of fitted model
}
