% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pglmm.compare.R
\name{print.pglmm.compare}
\alias{print.pglmm.compare}
\title{Print summary information of fitted model}
\usage{
\method{print}{pglmm.compare}(x, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{x}{A fitted pglmm.compare.}

\item{digits}{Minimal number of significant digits for printing, as in \code{\link{print.default}}.}

\item{...}{Additional arguments, currently ignored.}
}
\description{
Print summary information of fitted model
}
