% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pglmm_compare.R
\name{print.pglmm_compare}
\alias{print.pglmm_compare}
\title{Print summary information of fitted model}
\usage{
\method{print}{pglmm_compare}(x, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{x}{A fitted pglmm_compare.}

\item{digits}{Minimal number of significant digits for printing, as in \code{\link{print.default}}.}

\item{...}{Additional arguments, currently ignored.}
}
\description{
Print summary information of fitted model
}
