% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pglmm-utils.R
\docType{methods}
\name{ranef}
\alias{ranef}
\alias{ranef.communityPGLMM}
\alias{random.effects}
\title{Extract random-effects estimates}
\usage{
\method{ranef}{communityPGLMM}(object, ...)
}
\arguments{
\item{object}{A fitted model with class communityPGLMM.}

\item{...}{Ignored.}
}
\value{
A dataframe of random-effects estimates.
}
\description{
Extract the random-effects estimates
}
\details{
Extract the estimates of the random-effects parameters from a fitted model.
}
