% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pglmm-utils.R
\name{residuals.communityPGLMM}
\alias{residuals.communityPGLMM}
\title{Residuals of communityPGLMM objects}
\usage{
\method{residuals}{communityPGLMM}(
  object,
  type = if (object$family \%in\% c("binomial", "poisson")) "deviance" else "response",
  scaled = FALSE,
  ...
)
}
\arguments{
\item{object}{A fitted model with class communityPGLMM.}

\item{type}{Type of residuals, currently only "response" for gaussian pglmm;
"deviance" (default) and "response" for binomial and poisson pglmm.}

\item{scaled}{Scale residuals by residual standard deviation for gaussian pglmm.}

\item{\dots}{Additional arguments, ignored for method compatibility.}
}
\value{
A vector of residuals.
}
\description{
Getting different types of residuals for communityPGLMM objects.
}
