% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pglmm-utils.R
\name{simulate.communityPGLMM}
\alias{simulate.communityPGLMM}
\title{Simulate from a communityPGLMM object}
\usage{
\method{simulate}{communityPGLMM}(x, nsim = 1, seed = NULL, use.u = TRUE, ...)
}
\arguments{
\item{nsim}{number of response vectors to simulate.  Defaults to \code{1}.}

\item{seed}{an object specifying if and how the random number
    generator should be initialized (\sQuote{seeded}).\cr
    For the "lm" method, either \code{NULL} or an integer that will be
    used in a call to \code{set.seed} before simulating the response
    vectors.  If set, the value is saved as the \code{"seed"} attribute
    of the returned value.  The default, \code{NULL} will not change the
    random generator state, and return \code{\link{.Random.seed}} as the
    \code{"seed"} attribute, see \sQuote{Value}.
  }

\item{...}{additional optional arguments.}
}
\value{

}
\description{
Note that this function currently only works for model fit with \code{bayes = TRUE}
}
