% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pglmm-utils.R
\name{summary.communityPGLMM}
\alias{summary.communityPGLMM}
\title{Summary information of fitted model}
\usage{
\method{summary}{communityPGLMM}(object, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{object}{A fitted model with class communityPGLMM.}

\item{digits}{Minimal number of significant digits for printing, as in \code{\link{print.default}}.}

\item{...}{Additional arguments, currently ignored.}
}
\description{
Summary information of fitted model
}
