% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pglmm.compare.R
\name{summary.pglmm.compare}
\alias{summary.pglmm.compare}
\title{Summary information of fitted pglmm.compare model}
\usage{
\method{summary}{pglmm.compare}(object, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{object}{A fitted model with class pglmm.compare.}

\item{digits}{Minimal number of significant digits for printing, as in \code{\link{print.default}}.}

\item{...}{Additional arguments, currently ignored.}
}
\description{
Summary information of fitted pglmm.compare model
}
