% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pglmm_compare.R
\name{summary.pglmm_compare}
\alias{summary.pglmm_compare}
\title{Summary information of fitted pglmm_compare model}
\usage{
\method{summary}{pglmm_compare}(object, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{object}{A fitted model with class pglmm_compare.}

\item{digits}{Minimal number of significant digits for printing, as in \code{\link{print.default}}.}

\item{...}{Additional arguments, currently ignored.}
}
\description{
Summary information of fitted pglmm_compare model
}
