% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{test_nlopt}
\alias{test_nlopt}
\title{Testing optimization using nlopt: least squares for a simple regression}
\usage{
test_nlopt(xy, b0, b1, max_iter, method)
}
\arguments{
\item{xy}{A two-column matrix with the X variable as the first column and the Y
variable as the second one.}

\item{b0}{First estimate of the beta0 regression estimate.}

\item{b1}{First estimate of the beta1 regression estimate.}

\item{max_iter}{Maximum calls to the optimization function.}

\item{method}{Algorithm used for optimization. For now, the options are
"bobyqa", "cobyla", or "praxis".
See \url{https://nlopt.readthedocs.io/en/latest/NLopt_Algorithms}
for more information on these algorithms.}
}
\value{
A numeric vector with best set of parameters (b0 then b1) and
sum of squares corresponding to that set.
}
\description{
Testing optimization using nlopt: least squares for a simple regression
}
