% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{traits}
\alias{traits}
\title{Example species traits data}
\format{
A data frame with 18 species and 3 variables: sla,
height, and seed dispersal mode.
}
\usage{
traits
}
\description{
A data frame of species functional traits.
}
\keyword{datasets}
