% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{vcv2}
\alias{vcv2}
\title{Create phylogenetic var-cov matrix}
\usage{
vcv2(phy, corr = FALSE)
}
\arguments{
\item{phy}{a phylogeny with "phylo" as class.}

\item{corr}{whether to return a correlation matrix instead of Var-cov matrix. Default is FALSE}
}
\value{
a phylogenetic var-cov matrix
}
\description{
This function will convert a phylogeny to a Var-cov matrix.
}
