% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pglmm-utils.R
\name{fitted.communityPGLMM}
\alias{fitted.communityPGLMM}
\title{Fitted values for communityPGLMM}
\usage{
\method{fitted}{communityPGLMM}(object, ...)
}
\arguments{
\item{object}{A fitted model with class communityPGLMM.}

\item{\dots}{Additional arguments, ignored for method compatibility.}
}
\value{
Fitted values. For binomial and poisson PGLMMs, this is equal to mu.
}
\description{
Fitted values for communityPGLMM
}
