% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pglmm-utils.R
\name{get_design_matrix}
\alias{get_design_matrix}
\title{\code{get_design_matrix} gets design matrix for both gaussian and binomial models}
\usage{
get_design_matrix(formula, data, na.action = NULL, sp, site, random.effects)
}
\arguments{
\item{formula}{A two-sided linear formula object describing the
mixed-effects of the model; it follows similar syntax with \code{\link[lme4:lmer]{lmer}}.
There are some differences though.

First, to specify that a random term should have phylogenetic cov matrix along
with non-phylogenetic one, add \code{__} (two underscores) at the end of the group variable,
e.g. \code{+ (1 | sp__)} will construct two random terms,
one with phylogenetic cov matrix and another with non-phylogenetic (Identity) matrix;
However, \code{__} in the nested terms (below) will only create a phlylogenetic cov-matrix.
Therefore, nested random term has four forms:
\enumerate{
\item \code{(1|sp__@site)} represents correlated species are nested within independent sites
(i.e. kronecker(I_sites, V_sp)). This should be the most common one for community analysis.
\item \code{(1|sp@site__)} represents independent species are nested within correlated sites
(i.e. kron(V_sites, I_sp)). This one can be used for bipartite questions.
You can, for example, treat sp as insects and site as plants.
Remember to set the argument \code{tree_site} to a phylogeny.
\item \code{(1|sp__@site__)} represents correlated species are nested within correlated sites
(i.e. kron(V_sites, V_sp)). This one can also be used for bipartite questions.
\item \code{(1|sp@site)} will generate a identity matrix, which will be the same as
the residual of LMM. So not meaningful.
}

Second, note that correlated random terms will not be allowed at this moment. For example,
\code{(x|g)} will be equal with \code{(0 + x|g)} in the \code{lme4::lmer} syntax;
also, \code{(x1 + x2|g)} won't work.}

\item{data}{A \code{\link{data.frame}} containing the variables
named in formula. The data frame should have long format with
a column for species (named as 'sp') and a column for sites (named as 'site').
\code{communityPGLMM} will reorder rows of the data frame so that species
are nested within sites (i.e. arrange first by column site then by column sp).}

\item{na.action}{What to do with NAs?}

\item{sp}{No longer used, keep here for compatibility}

\item{site}{No longer used, keep here for compatibility}

\item{random.effects}{Pre-build list of random effects. If \code{NULL} (the default),
the function \code{\link{prep_dat_pglmm}} will prepare it for you. A list of pre-generated
random terms is also accepted (mainly to be compatible with code from previous versions).
If so, make sure that the orders of sp and site in the generated list are the same as the
data. This argument can be useful if users want to use other correlations (e.g. spatial)
in their models instead of phylogenetic relationships.}
}
\value{
A list of design matrices.
}
\description{
\code{get_design_matrix} gets design matrix for both gaussian and binomial models
}
