% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pglmm-utils.R
\name{predict.communityPGLMM}
\alias{predict.communityPGLMM}
\title{Predict Function for communityPGLMM Model Objects}
\usage{
\method{predict}{communityPGLMM}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{Object of class inheriting from \code{"lm"}}

\item{newdata}{An optional data frame in which to look for variables with
    which to predict.  If omitted, the fitted values are used.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
The form of the value returned by \code{predict} depends on the
  class of its argument.  See the documentation of the
  particular methods for details of what is produced by that method.
}
\description{
Predict Function for communityPGLMM Model Objects
}
