% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pglmm-utils.R
\name{simulate.communityPGLMM}
\alias{simulate.communityPGLMM}
\title{Simulate from a communityPGLMM object}
\usage{
\method{simulate}{communityPGLMM}(object, nsim = 1, seed = NULL, use.u = FALSE, ...)
}
\arguments{
\item{object}{an object representing a fitted model.}

\item{nsim}{number of response vectors to simulate.  Defaults to \code{1}.}

\item{seed}{an object specifying if and how the random number
    generator should be initialized (\sQuote{seeded}).\cr
    For the "lm" method, either \code{NULL} or an integer that will be
    used in a call to \code{set.seed} before simulating the response
    vectors.  If set, the value is saved as the \code{"seed"} attribute
    of the returned value.  The default, \code{NULL} will not change the
    random generator state, and return \code{\link{.Random.seed}} as the
    \code{"seed"} attribute, see \sQuote{Value}.
  }

\item{use.u}{(logical) if \code{TRUE}, generate a simulation
    conditional on the current random-effects estimates; if \code{FALSE}
    generate new Normally distributed random-effects values. (Redundant
    with \code{re.form}, which is preferred: \code{TRUE} corresponds to
    \code{re.form = NULL} (condition on all random effects), while
    \code{FALSE} corresponds to \code{re.form = ~0} (condition on none
    of the random effects).)}

\item{...}{additional optional arguments.}
}
\description{
Note that this function currently only works for model fit with \code{bayes = TRUE}
}
